Refactorizacion de codigo
