+++ apps/employes-api/src/app/empleados/empleados.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Post, Body, Get, Param, Delete, Patch } from '@nestjs/common';
-import { UpdatePersonaDto } from './dto/update_persona.dto';
-import { CreatePersonaDto } from './dto/create_persona.dto';
+import { UpdatePersonaDto } from './dto/update.persona.dto';
+import { CreatePersonaDto } from './dto/create.persona.dto';
 import { EmpleadosService } from './empleados.service';
 import { personal } from './empleados.entity';
