+++ apps/employes-api/src/main.ts
@@ -15,7 +15,7 @@ async function bootstrap() {
     optionsSuccessStatus: 204,
     credentials: true,
   });
-  
+
   await app.listen(port);
   Logger.log(
     `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`
